+++ test/examples/sample16-solution-with-solution-folders/test/proj1.tests/Library.fs
@@ -1,0 +1,5 @@
+﻿namespace proj1.tests
+
+module Say =
+    let hello name =
+        printfn "Hello %s" name